Optimize simulation pipeline for large N (2.1x speedup at N=100K)
Monolithic Numba Metropolis kernel with in-place Fisher-Yates shuffle
eliminates ~16 Python-Numba transitions per generation. Pre-allocated
pedigree arrays replace pd.concat accumulation. Vectorized dedup helper
replaces Python set loops. Fused midparent Numba kernel avoids ~50MB
temporaries per generation. Also replaces scipy rankdata/ndtri with
fast argsort-based rank and shared _ndtri_approx Numba kernel, and
vectorizes _subtract_pairs in pedigree_graph.py with np.isin.

+++ sim_ace/dropout.py
@@ -36,7 +36,7 @@ def run_dropout(pedigree: pd.DataFrame, params: dict) -> pd.DataFrame:
         return pedigree
 
     n_total = len(pedigree)
-    n_drop = int(round(n_total * rate))
+    n_drop = round(n_total * rate)
 
     if n_drop <= 0:
         logger.info("Dropout pass-through: n_drop=0 for rate=%.4f, N=%d", rate, n_total)
